% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicebar.R
\name{bcb}
\alias{bcb}
\title{Plot a lovely horizontal bar chart}
\usage{
bcb(
  df,
  var,
  group_var = NULL,
  title = NULL,
  x_label = NULL,
  y_label = NULL,
  group = FALSE,
  fill = NA,
  slice = TRUE,
  slice_n = 10,
  nudge_lab = -3,
  wrap = 26,
  xrange = NULL,
  sort_freq = TRUE
)
}
\arguments{
\item{df}{A data frame}

\item{var}{the main variable to plot}

\item{group_var}{the grouping variable (optional)}

\item{title}{An optional title as a string}

\item{x_label}{Optional label for the x axis}

\item{y_label}{Optional label for the y axis}

\item{group}{Logical variable marked FALSE by default. To group the bars this must be set to TRUE}

\item{fill}{Optional fill for bar chart}

\item{slice}{logical, if plot should include a sample of the categories}

\item{slice_n}{optional number of top categories sliced}

\item{nudge_lab}{adjust position of label}

\item{xrange}{optional range of x axis}

\item{sort_freq}{logical if data should be sorted by percentage}
}
\description{
Use this to make indiviudal or grouped bar charts with labels
}
