% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{bccmha_palettes_discrete}
\alias{bccmha_palettes_discrete}
\title{BCCMHA Palettes}
\format{
## `bccmha_palettes_discrete`
An object of class `palettes_palette` with
`r length(bccmha_palettes_discrete)` discrete colour palettes.
 Use `names(bccmha_palettes_discrete)` to return all palette names.
}
\usage{
bccmha_palettes_discrete
}
\description{
BCCMHA Palettes
}
\keyword{datasets}
