% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicebar.R
\name{bchist}
\alias{bchist}
\title{Plot a formatted histogram}
\usage{
bchist(
  df,
  var,
  group_var = NULL,
  group = FALSE,
  binwidth = NULL,
  title = NULL,
  x_label = NULL,
  y_label = NULL,
  fill = "#3179b4",
  color = "black",
  alpha = 0.8,
  wrap = 26,
  xrange = NULL,
  yrange = NULL
)
}
\arguments{
\item{df}{A data frame}

\item{var}{the main variable to plot}

\item{group_var}{the grouping variable (optional)}

\item{group}{Logical variable marked FALSE by default. To group the bars this must be set to TRUE}
}
\description{
Use this to make individual or grouped histograms
}
