% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicebar.R
\name{nicebar}
\alias{nicebar}
\title{Plot a lovely horizontal bar chart}
\usage{
nicebar(df, var, group_var = NULL, title = NULL, x_label = NULL, group = F)
}
\arguments{
\item{df}{A data frame}

\item{var}{the main variable to plot}

\item{group_var}{the grouping variable (optional)}

\item{title}{An optional title as a string}

\item{x_label}{Optional label for the x axis}

\item{group}{Logical variable marked FALSE by default. To group the bars this must be set to TRUE}
}
\description{
Use this to make indiviudal or grouped bar charts with labels
}
